Move mutators to separate package
