+++ internal/extension/manager.go
@@ -27,13 +27,10 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types/ref"
-	authorinov1beta3 "github.com/kuadrant/authorino/api/v1beta3"
-	"github.com/kuadrant/policy-machinery/machinery"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
-	kuadrantv1 "github.com/kuadrant/kuadrant-operator/api/v1"
 	kuadrant "github.com/kuadrant/kuadrant-operator/pkg/cel/ext"
 	extpb "github.com/kuadrant/kuadrant-operator/pkg/extension/grpc/v1"
 )
@@ -44,41 +41,6 @@ type Manager struct {
 	dag        *nilGuardedPointer[StateAwareDAG]
 	logger     logr.Logger
 	sync       io.Writer
-}
-
-type ResourceMutator[TResource any, TPolicy machinery.Policy] interface {
-	Mutate(resource TResource, policy TPolicy) error
-}
-
-type AuthConfigMutator = ResourceMutator[*authorinov1beta3.AuthConfig, *kuadrantv1.AuthPolicy]
-
-type MutatorRegistry struct {
-	authConfigMutators []AuthConfigMutator
-	mutex              sync.RWMutex
-}
-
-var GlobalMutatorRegistry = &MutatorRegistry{}
-
-func (r *MutatorRegistry) RegisterAuthConfigMutator(mutator AuthConfigMutator) {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-	r.authConfigMutators = append(r.authConfigMutators, mutator)
-}
-
-func (r *MutatorRegistry) ApplyAuthConfigMutators(authConfig *authorinov1beta3.AuthConfig, policy *kuadrantv1.AuthPolicy) error {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-
-	for _, mutator := range r.authConfigMutators {
-		if err := mutator.Mutate(authConfig, policy); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func ApplyAuthConfigMutators(authConfig *authorinov1beta3.AuthConfig, policy *kuadrantv1.AuthPolicy) error {
-	return GlobalMutatorRegistry.ApplyAuthConfigMutators(authConfig, policy)
 }
 
 type Extension interface {
@@ -164,64 +126,6 @@ func (m *Manager) HasSynced() bool {
 	return true
 }
 
-type RegisteredDataStore struct {
-	data  map[string]map[string]RegisteredDataEntry
-	mutex sync.RWMutex
-}
-
-func NewRegisteredDataStore() *RegisteredDataStore {
-	return &RegisteredDataStore{
-		data: make(map[string]map[string]RegisteredDataEntry),
-	}
-}
-
-func (r *RegisteredDataStore) Set(target, requester, binding string, entry RegisteredDataEntry) {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	entryKey := fmt.Sprintf("%s#%s", requester, binding)
-
-	if _, exists := r.data[target]; !exists {
-		r.data[target] = make(map[string]RegisteredDataEntry)
-	}
-
-	r.data[target][entryKey] = entry
-}
-
-func (r *RegisteredDataStore) GetAllForTarget(target string) []RegisteredDataEntry {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-
-	entries, exists := r.data[target]
-	if !exists || len(entries) == 0 {
-		return nil
-	}
-
-	result := make([]RegisteredDataEntry, 0, len(entries))
-	for _, entry := range entries {
-		result = append(result, entry)
-	}
-	return result
-}
-
-func (r *RegisteredDataStore) Delete(target, requester, binding string) bool {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
-
-	entryKey := fmt.Sprintf("%s#%s", requester, binding)
-
-	if targetMap, exists := r.data[target]; exists {
-		if _, entryExists := targetMap[entryKey]; entryExists {
-			delete(targetMap, entryKey)
-			if len(targetMap) == 0 {
-				delete(r.data, target)
-			}
-			return true
-		}
-	}
-	return false
-}
-
 type extensionService struct {
 	dag            *nilGuardedPointer[StateAwareDAG]
 	mutex          sync.Mutex
@@ -243,9 +147,7 @@ func newExtensionService(dag *nilGuardedPointer[StateAwareDAG]) extpb.ExtensionS
 		registeredData: NewRegisteredDataStore(),
 	}
 
-	mutator := &RegisteredDataMutator{
-		registeredData: service.registeredData,
-	}
+	mutator := NewRegisteredDataMutator(service.registeredData)
 	GlobalMutatorRegistry.RegisterAuthConfigMutator(mutator)
 
 	return service
@@ -346,55 +248,6 @@ type subscription struct {
 	cAst  *cel.Ast
 	input map[string]any
 	val   ref.Val
-}
-
-type RegisteredDataMutator struct {
-	registeredData *RegisteredDataStore
-}
-
-func (m *RegisteredDataMutator) Mutate(authConfig *authorinov1beta3.AuthConfig, policy *kuadrantv1.AuthPolicy) error {
-	policyKey := fmt.Sprintf("%s/%s/%s", policy.GetObjectKind().GroupVersionKind().Kind, policy.GetNamespace(), policy.GetName())
-
-	registeredEntries := m.registeredData.GetAllForTarget(policyKey)
-	if len(registeredEntries) == 0 {
-		return nil
-	}
-
-	if authConfig.Spec.Response == nil {
-		authConfig.Spec.Response = &authorinov1beta3.ResponseSpec{
-			Success: authorinov1beta3.WrappedSuccessResponseSpec{
-				DynamicMetadata: make(map[string]authorinov1beta3.SuccessResponseSpec),
-			},
-		}
-	} else if authConfig.Spec.Response.Success.DynamicMetadata == nil {
-		authConfig.Spec.Response.Success.DynamicMetadata = make(map[string]authorinov1beta3.SuccessResponseSpec)
-	}
-
-	// Collect all properties from all entries
-	properties := make(map[string]authorinov1beta3.ValueOrSelector)
-	for _, entry := range registeredEntries {
-		properties[entry.Binding] = authorinov1beta3.ValueOrSelector{
-			Expression: authorinov1beta3.CelExpression(entry.Expression),
-		}
-	}
-
-	// Set all properties at once
-	authConfig.Spec.Response.Success.DynamicMetadata["kuadrant"] = authorinov1beta3.SuccessResponseSpec{
-		AuthResponseMethodSpec: authorinov1beta3.AuthResponseMethodSpec{
-			Json: &authorinov1beta3.JsonAuthResponseSpec{
-				Properties: properties,
-			},
-		},
-	}
-
-	return nil
-}
-
-type RegisteredDataEntry struct {
-	Requester  string
-	Binding    string
-	Expression string
-	CAst       *cel.Ast
 }
 
 func (s *extensionService) RegisterMutator(_ context.Context, request *extpb.RegisterMutatorRequest) (*emptypb.Empty, error) {
